third_party/WebKit/LayoutTests/typedcssom/usecounters/computed-stylePropertyMap.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<div id="target">
<script>
'use strict';

test(() => {
  const CSSTypedOMStylePropertyMap = 2334; // From UseCounter.h

  assert_false(internals.isUseCounted(document, CSSTypedOMStylePropertyMap));

  document.getElementById('target').computedStyleMap();

  assert_true(internals.isUseCounted(document, CSSTypedOMStylePropertyMap));
}, 'We UseCount when we obtain a computed style property map');

</script>
